Reorganize src folders
